# openapi.yaml
openapi: 3.1.0
info:
  title: Hub API
  description: >
    API contract for the Hub backend (FastAPI). Derived from the frontend's
    API client (frontend/src/services/api/) and the data model in
    docs/specs.md. The backend implements this contract;
    backend/tests/test_openapi_contract.py fails if the paths and methods
    here and in the FastAPI app drift apart.

    Any operation that reads or writes the database can also answer 503
    with a Retry-After header and an HTTPError body when the database
    can't be reached (for example while Neon resumes from suspend).
    Clients should retry; it never means the session is invalid.
  version: 0.1.0
servers:
  - url: https://hub.dnls.dev/api
    description: Production (CloudFront forwards /api/* to the API)
  - url: https://hub-dev.dnls.dev/api
    description: Development (CloudFront forwards /api/* to the API)
  - url: http://localhost:8000
    description: Local development

tags:
  - name: auth
  - name: projects
  - name: notes
  - name: health
  - name: errors

security:
  - bearerAuth: []
  - cookieAuth: []

paths:
  /health:
    get:
      tags: [health]
      summary: Liveness check
      description: Unauthenticated. Does not touch the database.
      security: []
      responses:
        '200':
          description: The API is up
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    enum: [ok]

  /auth/register:
    post:
      tags: [auth]
      summary: Self-register a new account
      description: >
        Limited to REGISTER_RATE_LIMIT_PER_MINUTE attempts per client IP per
        minute (3 when deployed); over it, 429 with Retry-After.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          description: >
            Account created and logged in. The token is in the body and also
            set as the httpOnly `hub_token` session cookie.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '403':
          description: >
            Sign-ups are closed: the instance has MAX_ACCOUNTS accounts (1000
            when deployed; off locally).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          description: Too many sign-up attempts from this client IP
          headers:
            Retry-After:
              description: Seconds until another attempt is allowed
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'

  /auth/login:
    post:
      tags: [auth]
      summary: Log in (OAuth2 password flow)
      description: >
        Standard OAuth2 password grant so FastAPI's built-in OAuth2PasswordBearer
        / Swagger "Authorize" tooling works out of the box. The email goes in the
        `username` field. Limited to LOGIN_RATE_LIMIT_PER_MINUTE attempts per
        client IP per minute (5 when deployed); over it, 429 with Retry-After.
        A password over 256 characters (the sign-up maximum) is a 422 before
        any hashing.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginForm'
      responses:
        '200':
          description: >
            Authenticated. The token is in the body and also set as the
            httpOnly `hub_token` session cookie.
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Incorrect email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          description: Too many login attempts from this client IP
          headers:
            Retry-After:
              description: Seconds until another attempt is allowed
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'

  /auth/logout:
    post:
      tags: [auth]
      summary: Clear the session cookie
      description: >
        Expires the `hub_token` cookie. Needs no credentials. Tokens are
        stateless JWTs, so a token held elsewhere (e.g. a bearer token an API
        client kept) stays valid until it expires.
      security: []
      responses:
        '204':
          description: Cookie cleared
          headers:
            Set-Cookie:
              $ref: '#/components/headers/SessionCookie'

  /auth/me:
    get:
      tags: [auth]
      summary: Get the current authenticated user
      responses:
        '200':
          description: The current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /projects:
    get:
      tags: [projects]
      summary: List the current user's projects
      description: >
        Returns all of the authenticated user's projects, unsorted and
        unfiltered. Filtering by status/tag and sorting by score or date
        (per the spec's dashboard) is done client-side today; the frontend's
        listProjects() call takes no parameters.
      responses:
        '200':
          description: The user's projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectInput'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: >
            The user already has MAX_PROJECTS_PER_USER projects (default 500).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '422':
          $ref: '#/components/responses/ValidationError'

  /projects/{project_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [projects]
      summary: Get a single project
      responses:
        '200':
          description: The project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [projects]
      summary: Update a project
      description: >
        Partial update. Any subset of fields may be sent. `updatedAt` is
        always server-set to the current time on a successful update and
        cannot be set by the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectInput'
      responses:
        '200':
          description: The updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      tags: [projects]
      summary: Delete a project
      description: Also deletes the project's notes (cascade).
      responses:
        '204':
          description: Project deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /projects/{project_id}/notes:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [notes]
      summary: List a project's notes
      description: Returned newest first (descending by createdAt).
      responses:
        '200':
          description: The project's notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Note'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [notes]
      summary: Add a note to a project
      description: >
        Adding a note also bumps the parent project's updatedAt, so the
        response includes both the created note and the updated project.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNoteInput'
      responses:
        '201':
          description: Note created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddNoteResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: >
            The project already has MAX_NOTES_PER_PROJECT notes (default 500).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

  /notes/{note_id}:
    parameters:
      - $ref: '#/components/parameters/NoteId'
    delete:
      tags: [notes]
      summary: Delete a note
      description: >
        Ownership is enforced via the parent project (notes have no owner of
        their own). Also bumps the parent project's updatedAt, so the
        response is the updated parent project.
      responses:
        '200':
          description: The note's parent project, updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /client-errors:
    post:
      tags: [errors]
      summary: Report a frontend error
      description: >
        The web app reports uncaught errors, unhandled promise rejections,
        render crashes and failed API calls (network errors and 5xx) here.
        Each report is written as one JSON log line; nothing is stored in the
        database. Unauthenticated, since errors happen while signed out too;
        when a valid session is present the user id is attached. URLs are
        logged without their query string or fragment. Limited to
        CLIENT_ERROR_RATE_LIMIT_PER_MINUTE reports per client IP per minute
        (30 when deployed); over it, 429 with Retry-After.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientErrorReport'
      responses:
        '204':
          description: Report logged
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          description: Too many error reports from this client IP
          headers:
            Retry-After:
              description: Seconds until another report is allowed
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: hub_token
      description: >
        The web app's session: the same JWT, in an httpOnly, SameSite=Strict
        cookie (Secure when deployed) set by /auth/login and /auth/register.
        When a request sends both, the Authorization header wins.

  headers:
    SessionCookie:
      description: >
        `hub_token=<JWT>; HttpOnly; SameSite=Strict; Path=/; Secure` (Secure
        when deployed), or an expired `hub_token` to clear it.
      schema:
        type: string

  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    NoteId:
      name: note_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPError'
    NotFound:
      description: >
        Resource does not exist, or exists but is owned by a different user.
        The two cases are indistinguishable by design, to avoid leaking
        existence of other users' data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPError'
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'

  schemas:
    Status:
      type: string
      enum: [Inbox, Exploring, Active, Parked, Graduated, Killed]

    User:
      type: object
      required: [id, email, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: [string, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    RegisterInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 256
        displayName:
          type: [string, "null"]

    LoginForm:
      type: object
      description: OAuth2 password grant fields (application/x-www-form-urlencoded).
      required: [username, password]
      properties:
        grant_type:
          type: string
          enum: [password]
        username:
          type: string
          format: email
          description: The account's email address.
        password:
          type: string
          maxLength: 256
        scope:
          type: string
        client_id:
          type: string
        client_secret:
          type: string

    TokenResponse:
      type: object
      description: >
        Field names follow the OAuth2 spec (snake_case), not the camelCase
        convention used elsewhere in this API, for compatibility with
        standard OAuth2 client tooling.
      required: [access_token, token_type]
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [bearer]

    Project:
      type: object
      required:
        - id
        - name
        - pitch
        - description
        - status
        - tags
        - excitement
        - effort
        - potential
        - nextAction
        - targetDate
        - links
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        pitch:
          type: string
          description: One-line, scannable pitch, separate from description.
        description:
          type: string
          description: The full brain-dump.
        status:
          $ref: '#/components/schemas/Status'
        tags:
          type: array
          items:
            type: string
        excitement:
          type: integer
          minimum: 1
          maximum: 5
        effort:
          type: integer
          minimum: 1
          maximum: 5
        potential:
          type: integer
          minimum: 1
          maximum: 5
        nextAction:
          type: string
          description: The single next concrete step (GTD core).
        targetDate:
          type: [string, "null"]
          format: date
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateProjectInput:
      type: object
      required: [name]
      description: >
        Text fields are capped in characters (not bytes); longer values are a
        422. The database enforces the same limits.
      properties:
        name:
          type: string
          maxLength: 256
        pitch:
          type: string
          maxLength: 2000
        description:
          type: string
          maxLength: 5000
        status:
          $ref: '#/components/schemas/Status'
        tags:
          type: array
          items:
            type: string
        excitement:
          type: integer
          minimum: 1
          maximum: 5
        effort:
          type: integer
          minimum: 1
          maximum: 5
        potential:
          type: integer
          minimum: 1
          maximum: 5
        nextAction:
          type: string
          maxLength: 1000
        targetDate:
          type: [string, "null"]
          format: date
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'

    UpdateProjectInput:
      type: object
      description: >
        Partial update; every field optional. Excludes id, createdAt, and
        updatedAt, which are never client-settable. Same text limits as
        CreateProjectInput.
      properties:
        name:
          type: string
          maxLength: 256
        pitch:
          type: string
          maxLength: 2000
        description:
          type: string
          maxLength: 5000
        status:
          $ref: '#/components/schemas/Status'
        tags:
          type: array
          items:
            type: string
        excitement:
          type: integer
          minimum: 1
          maximum: 5
        effort:
          type: integer
          minimum: 1
          maximum: 5
        potential:
          type: integer
          minimum: 1
          maximum: 5
        nextAction:
          type: string
          maxLength: 1000
        targetDate:
          type: [string, "null"]
          format: date
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'

    Link:
      type: object
      required: [url]
      properties:
        label:
          type: [string, "null"]
          description: Optional. Returned as null when not set.
        url:
          type: string
          format: uri
          pattern: '^[Hh][Tt][Tt][Pp][Ss]?://'
          description: Must start with http:// or https://.

    Note:
      type: object
      required: [id, projectId, body, createdAt]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        body:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateNoteInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          maxLength: 10000
          description: Capped in characters; longer is a 422.

    AddNoteResponse:
      type: object
      required: [note, project]
      properties:
        note:
          $ref: '#/components/schemas/Note'
        project:
          $ref: '#/components/schemas/Project'

    ClientErrorReport:
      type: object
      required: [kind, message]
      properties:
        kind:
          type: string
          enum: [error, unhandledrejection, render, http]
        message:
          type: string
          maxLength: 2000
        stack:
          type: [string, "null"]
          maxLength: 8000
        url:
          type: [string, "null"]
          maxLength: 2000
          description: The page the error happened on.
        userAgent:
          type: [string, "null"]
          maxLength: 500
        endpoint:
          type: [string, "null"]
          maxLength: 500
          description: For kind http, the API URL that failed.
        method:
          type: [string, "null"]
          maxLength: 10
        status:
          type: [integer, "null"]
          minimum: 0
          maximum: 599
          description: For kind http. 0 means no response (network error).

    HTTPError:
      type: object
      required: [detail]
      properties:
        detail:
          type: string

    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  anyOf:
                    - type: string
                    - type: integer
              msg:
                type: string
              type:
                type: string